Core update 8.8.5, webform update 5.11 and composer updates

+++ profiles/express/modules/contrib/webform/src/Form/AdminConfig/WebformAdminConfigAdvancedForm.php
@@ -8,6 +8,7 @@ use Drupal\Core\Extension\ModuleHandlerInterface;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Routing\RouteBuilderInterface;
 use Drupal\Core\Url;
+use Drupal\webform\Commands\WebformCliService;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
@@ -37,6 +38,13 @@ class WebformAdminConfigAdvancedForm extends WebformAdminConfigBaseForm {
   protected $routerBuilder;
 
   /**
+   * The (drush) command-line service.
+   *
+   * @var \Drupal\webform\Commands\WebformCliService
+   */
+  protected $cliService;
+
+  /**
    * {@inheritdoc}
    */
   public function getFormId() {
@@ -52,14 +60,17 @@ class WebformAdminConfigAdvancedForm extends WebformAdminConfigBaseForm {
    *   The module handler.
    * @param \Drupal\Core\Cache\CacheBackendInterface $render_cache
    *   The render cache service.
-   * @param \Drupal\Core\Routing\RouteBuilderInterface $route_builder
+   * @param \Drupal\Core\Routing\RouteBuilderInterface $router_builder
    *   The router builder service.
+   * @param \Drupal\webform\Commands\WebformCliService $cli_service
+   *   The (drush) command-line service.
    */
-  public function __construct(ConfigFactoryInterface $config_factory, ModuleHandlerInterface $module_handler, CacheBackendInterface $render_cache, RouteBuilderInterface $router_builder) {
+  public function __construct(ConfigFactoryInterface $config_factory, ModuleHandlerInterface $module_handler, CacheBackendInterface $render_cache, RouteBuilderInterface $router_builder, WebformCliService $cli_service) {
     parent::__construct($config_factory);
-    $this->moduleHandler = $module_handler;
     $this->renderCache = $render_cache;
+    $this->moduleHandler = $module_handler;
     $this->routerBuilder = $router_builder;
+    $this->cliService = $cli_service;
   }
 
   /**
@@ -70,7 +81,8 @@ class WebformAdminConfigAdvancedForm extends WebformAdminConfigBaseForm {
       $container->get('config.factory'),
       $container->get('module_handler'),
       $container->get('cache.render'),
-      $container->get('router.builder')
+      $container->get('router.builder'),
+      $container->get('webform.cli_service')
     );
   }
 
@@ -115,27 +127,26 @@ class WebformAdminConfigAdvancedForm extends WebformAdminConfigBaseForm {
       '#return_value' => TRUE,
       '#default_value' => $config->get('ui.details_save'),
     ];
+    $form['ui']['help_disabled'] = [
+      '#type' => 'checkbox',
+      '#title' => $this->t('Disable help'),
+      '#description' => $this->t('If checked, help text will be removed from every webform page and form.'),
+      '#return_value' => TRUE,
+      '#default_value' => $config->get('ui.help_disabled'),
+    ];
     $form['ui']['dialog_disabled'] = [
       '#type' => 'checkbox',
       '#title' => $this->t('Disable dialogs'),
-      '#description' => $this->t('If checked, all modal dialogs (i.e. popups) will be disabled.'),
+      '#description' => $this->t('If checked, all modal/off-canvas dialogs (i.e. popups) will be disabled.'),
       '#return_value' => TRUE,
       '#default_value' => $config->get('ui.dialog_disabled'),
-    ];
-    $form['ui']['about_disabled'] = [
-      '#type' => 'checkbox',
-      '#title' => $this->t("Disable the 'About' section"),
-      '#description' => $this->t("If checked, 'About' section/tab will be remove from the admin UI."),
-      '#return_value' => TRUE,
-      '#default_value' => $config->get('ui.about_disabled'),
     ];
     $form['ui']['offcanvas_disabled'] = [
       '#type' => 'checkbox',
       '#title' => $this->t('Disable off-canvas system tray'),
-      '#description' => $this->t('If checked, the off-canvas system tray will be disabled.'),
+      '#description' => $this->t('If checked, all off-canvas system trays will be disabled.'),
       '#return_value' => TRUE,
       '#default_value' => $config->get('ui.offcanvas_disabled'),
-      '#access' => $this->moduleHandler->moduleExists('outside_in') && (floatval(\Drupal::VERSION) >= 8.3),
       '#states' => [
         'visible' => [
           ':input[name="ui[dialog_disabled]"]' => [
@@ -153,26 +164,11 @@ class WebformAdminConfigAdvancedForm extends WebformAdminConfigBaseForm {
       '#return_value' => TRUE,
       '#default_value' => $config->get('ui.promotions_disabled'),
     ];
-    if (!$this->moduleHandler->moduleExists('outside_in') && (floatval(\Drupal::VERSION) >= 8.3)) {
-      $form['ui']['offcanvas_message'] = [
-        '#type' => 'webform_message',
-        '#message_type' => 'info',
-        '#message_message' => $this->t('Enable the experimental <a href=":href">System tray module</a> to improve the Webform module\'s user experience.', [':href' => 'https://www.drupal.org/blog/drupal-82-now-with-more-outside-in']),
-        '#states' => [
-          'visible' => [
-            ':input[name="ui[dialog_disabled]"]' => [
-              'checked' => FALSE,
-            ],
-          ],
-        ],
-        '#weight' => -100,
-      ];
-    }
 
     // Requirements.
     $form['requirements'] = [
       '#type' => 'details',
-      '#title' => $this->t('Requirements'),
+      '#title' => $this->t('Requirement settings'),
       '#description' => $this->t('The below requirements are checked by the <a href=":href">Status report</a>.', [':href' => Url::fromRoute('system.status')->toString()]),
       '#open' => TRUE,
       '#tree' => TRUE,
@@ -180,7 +176,8 @@ class WebformAdminConfigAdvancedForm extends WebformAdminConfigBaseForm {
     $form['requirements']['cdn'] = [
       '#type' => 'checkbox',
       '#title' => $this->t('Check if CDN is being used for external libraries'),
-      '#description' => $this->t('If unchecked, all warnings about missing libraries will be disabled.'),
+      '#description' => $this->t('If unchecked, all warnings about missing libraries will be disabled.') . '<br/><br/>' .
+        $this->t('Relying on a CDN for external libraries can cause unexpected issues with Ajax and BigPipe support. For more information see: <a href=":href">Issue #1988968</a>', [':href' => 'https://www.drupal.org/project/drupal/issues/1988968']),
       '#return_value' => TRUE,
       '#default_value' => $config->get('requirements.cdn'),
     ];
@@ -216,8 +213,8 @@ class WebformAdminConfigAdvancedForm extends WebformAdminConfigBaseForm {
     $form['test']['names'] = [
       '#type' => 'webform_codemirror',
       '#mode' => 'yaml',
-      '#title' => $this->t('Test data by element name'),
-      '#description' => $this->t("Above test data is keyed by full or partial element names. For example, using 'zip' will populate fields that are named 'zip' and 'zip_code' but not 'zipcode' or 'zipline'."),
+      '#title' => $this->t('Test data by element key'),
+      '#description' => $this->t("Above test data is keyed by full or partial element keys. For example, using 'zip' will populate element keys that are 'zip' and 'zip_code' but not 'zipcode' or 'zipline'."),
       '#default_value' => $config->get('test.names'),
     ];
 
@@ -234,6 +231,15 @@ class WebformAdminConfigAdvancedForm extends WebformAdminConfigBaseForm {
       '#min' => 1,
       '#required' => TRUE,
       '#default_value' => $config->get('batch.default_batch_export_size'),
+      '#description' => $this->t('Batch export size is used when submissions are being exported/downloaded.'),
+    ];
+    $form['batch']['default_batch_import_size'] = [
+      '#type' => 'number',
+      '#title' => $this->t('Batch import size'),
+      '#min' => 1,
+      '#required' => TRUE,
+      '#default_value' => $config->get('batch.default_batch_import_size'),
+      '#description' => $this->t('Batch import size is used when submissions are being imported/uploaded.'),
     ];
     $form['batch']['default_batch_update_size'] = [
       '#type' => 'number',
@@ -241,10 +247,12 @@ class WebformAdminConfigAdvancedForm extends WebformAdminConfigBaseForm {
       '#min' => 1,
       '#required' => TRUE,
       '#default_value' => $config->get('batch.default_batch_update_size'),
+      '#description' => $this->t('Batch update size is used when submissions are being bulk updated.'),
     ];
     $form['batch']['default_batch_delete_size'] = [
       '#type' => 'number',
       '#title' => $this->t('Batch delete size'),
+      '#description' => $this->t('Batch delete size is used when submissions are being cleared.'),
       '#min' => 1,
       '#required' => TRUE,
       '#default_value' => $config->get('batch.default_batch_delete_size'),
@@ -258,6 +266,48 @@ class WebformAdminConfigAdvancedForm extends WebformAdminConfigBaseForm {
       '#default_value' => $config->get('batch.default_batch_email_size'),
     ];
 
+    // Repair.
+    $form['repair'] = [
+      '#type' => 'details',
+      '#title' => $this->t('Repair webform configuration'),
+      '#open' => TRUE,
+      '#help' => FALSE,
+      '#weight' => 100,
+    ];
+    $form['repair']['warning'] = [
+      '#type' => 'webform_message',
+      '#message_type' => 'warning',
+      '#message_message' => $this->t('Repair and remove older Webform configuration files.') . '<br/>' .
+        '<strong>' . $this->t('This action cannot be undone.') . '</strong>',
+    ];
+    $form['repair'] += [
+      'title' => [
+        '#markup' => $this->t('This action will…'),
+      ],
+      'list' => [
+        '#theme' => 'item_list',
+        '#items' => [
+          $this->t('Repair webform submission storage schema'),
+          $this->t('Repair admin configuration'),
+          $this->t('Repair webform settings'),
+          $this->t('Repair webform handlers'),
+          $this->t('Repair webform field storage definitions'),
+          $this->t('Repair webform submission storage schema'),
+          $this->t('Remove webform submission translation settings'),
+        ],
+      ],
+    ];
+    $form['repair']['action'] = ['#type' => 'actions'];
+    $form['repair']['action']['repair_configuration'] = [
+      '#type' => 'submit',
+      '#value' => $this->t('Repair configuration'),
+      '#attributes' => [
+        'onclick' => 'return confirm("' . $this->t('Are you sure you want to repair and remove older webform configuration?')
+          . PHP_EOL
+          . $this->t('This cannot be undone!!!') . '");',
+      ],
+    ];
+
     return parent::buildForm($form, $form_state);
   }
 
@@ -265,19 +315,71 @@ class WebformAdminConfigAdvancedForm extends WebformAdminConfigBaseForm {
    * {@inheritdoc}
    */
   public function submitForm(array &$form, FormStateInterface $form_state) {
-    $config = $this->config('webform.settings');
-    $config->set('ui', $form_state->getValue('ui'));
-    $config->set('requirements', $form_state->getValue('requirements'));
-    $config->set('test', $form_state->getValue('test'));
-    $config->set('batch', $form_state->getValue('batch'));
-    $config->save();
-    
-    // Clear render cache so that local tasks can be updated.
-    // @see webform_local_tasks_alter()
-    $this->renderCache->deleteAll();
-    \Drupal::service('router.builder')->rebuild();
+    $op = (string) $form_state->getValue('op');
+    if ($op === (string) $this->t('Repair configuration')) {
+      // Copied from:
+      // @see \Drupal\webform\Commands\WebformCliService::drush_webform_repair
+      module_load_include('install', 'webform');
 
-    parent::submitForm($form, $form_state);
+      $this->messenger()->addMessage($this->t('Repairing webform submission storage schema…'));
+      _webform_update_webform_submission_storage_schema();
+
+      $this->messenger()->addMessage($this->t('Repairing admin configuration…'));
+      _webform_update_admin_settings(TRUE);
+
+      $this->messenger()->addMessage($this->t('Repairing webform settings…'));
+      _webform_update_webform_settings();
+
+      $this->messenger()->addMessage($this->t('Repairing webform handlers…'));
+      _webform_update_webform_handler_settings();
+
+      $this->messenger()->addMessage($this->t('Repairing webform field storage definitions…'));
+      _webform_update_field_storage_definitions();
+
+      $this->messenger()->addMessage($this->t('Repairing webform submission storage schema…'));
+      _webform_update_webform_submission_storage_schema();
+
+      if ($this->moduleHandler->moduleExists('webform_entity_print')) {
+        $this->messenger()->addMessage($this->t('Repairing webform entity print settings…'));
+        module_load_include('install', 'webform_entity_print');
+        webform_entity_print_install();
+      }
+
+      $this->messenger()->addMessage($this->t('Removing (unneeded) webform submission translation settings…'));
+      _webform_update_webform_submission_translation();
+
+      drupal_flush_all_caches();
+
+      $this->messenger()->addStatus($this->t('Webform configuration has been repaired.'));
+    }
+    else {
+      // Update config and submit form.
+      $config = $this->config('webform.settings');
+      $config->set('ui', $form_state->getValue('ui'));
+      $config->set('requirements', $form_state->getValue('requirements'));
+      $config->set('test', $form_state->getValue('test'));
+      $config->set('batch', $form_state->getValue('batch'));
+
+      // Track if help is disabled.
+      // @todo Figure out how to clear cached help block.
+      $is_help_disabled = ($config->getOriginal('ui.help_disabled') != $config->get('ui.help_disabled'));
+
+      parent::submitForm($form, $form_state);
+
+      // Clear cached data.
+      if ($is_help_disabled) {
+        // Flush cache when help is being enabled.
+        // @see webform_help()
+        drupal_flush_all_caches();
+      }
+      else {
+        // Clear render cache so that local tasks can be updated to hide/show
+        // the 'Contribute' tab.
+        // @see webform_local_tasks_alter()
+        $this->renderCache->deleteAll();
+        $this->routerBuilder->rebuild();
+      }
+    }
   }
 
 }
